Create a loop that listens to stdin and a socket.
Internet sockets and input streams are typically blocking, but this is a
problem for programs that need to do both at once.  Instead of splitting the
tasks into two different threads, which can lead to its own challenges, I used
system calls to determine which file descriptors are available for reading and
writing.

This script is not finished in any sense.  However, it will respond to
keyboard input and incoming connections.  It listens to port 12345 on
localhost.